Simplest estimator for variance.

An estimator for the variance of a matrix of independent random variables when assuming the mean matrix is low rank with the rank known.

sig_naive(d, N, p, r)

Arguments

d
A vector of positive numerics. The singular values of the data matrix.

N
A positive integer. The row dimension.

p
A positive integer. The column dimension.

r
A positive integer. The known rank of the matrix. Must be less than min(N, p)

Value

sig2_est A positive numeric. The estimate of the variance.

Details

Let \(Y\) be a matrix with row dimension \(n\) and column dimension \(p\) where \(n \ge p\). The model is \(Y = \Theta + \sigma E\) where \(\Theta\) is low rank and \(E\) contains independent elements with mean zero and variance one. The MLE of \(\sigma^2\) is the sum of squares of the last \(p - r\) singular values divided by \(pn\). This estimator has negative bias so Choi et al (2014) suggested to instead divide by \(n(p - r)\). This is the estimator implemented here. It seems to have reasonable performance when the rank is small. Though the biggest drawback here is that you have to know the rank ahead of time.

References

Choi, Yunjin, Jonathan Taylor, and Robert Tibshirani. "Selecting the number of principal components: Estimation of the true rank of a noisy matrix." arXiv preprint arXiv:1410.8260 (2014).