Variance estimators from Choi et al (2014).

Returns a degrees of freedom corrected residual mean-squared error of a soft-thresholding estimator where the tuning parameter is chosen by a cross-validation-ish procedure.

sig_soft(Y, c_val = 2/3, k = 10, lambda_grid = NULL,
  print_update = FALSE)

Arguments

Y
The data matrix.

c_val
The ad-hoc adjustment to the degrees of freedom. Choi et al (2014) found that 2/3 worked well in simultations.

k
A positive integer. The fold for the soft-impute cross validation. Default is 10.

lambda_grid
A vector of positive numerics. The values of lambda to compute. The default is 20 values from the minimum to the maximum singular value of Y.

print_update
A logical. Should we print to the screen the status of the cross-validation-ish procedure at each iteration (TRUE) or not (FALSE)?

Value

sig2_est A positive numeric. The estimate of the variance.

sse A positive numeric. The sum of squared errors for estimated Y.

dfLambda A positive integer. The estimated df.

Details

See Choi et al (2014) for details. This seems to be the best estimation procedure so far, but also takes the longest.

You can try out different values of c_val with the outputs of sse and dfLambda

References

Choi, Yunjin, Jonathan Taylor, and Robert Tibshirani. "Selecting the number of principal components: Estimation of the true rank of a noisy matrix." arXiv preprint arXiv:1410.8260 (2014).

See also