Runs an iterative coordinate descent algorithm to minimize the SURE for mode-specific soft thresholding estimators.

Runs an iterative coordinate descent algorithm to minimize the SURE for mode-specific soft thresholding estimators.

soft_coord(c_obj, lambda_init, c_init, itermax = 1000, tol = 10^-4,
  print_iter = TRUE, tau2 = 1, use_sure = TRUE)

Arguments

c_obj
The output from get_c.

lambda_init
A vector of numerics of length \(n\). The initial starting values for the thresholding parameters.

c_init
A positive numeric. The starting value of the scaling parameter.

itermax
A positive integer. The maximum number of Newton steps to iterate through.

tol
A positive numeric. The stopping criterion.

print_iter
A logical. Should we print the updates of the Newton Step?

tau2
A positive numeric. The variance. Assumed known and defaults to 1.

use_sure
A logical. Which stopping criterion should we use? The mean absolute difference in the parameters (FALSE) or the absolute value of the deviation of the ratio of adjacent SURE values from 1 (TRUE).

Value

c A numeric. The final value of the scaling parameter.

lambda A vector of numerics. The final values of the thresholding parameters.

est An array of numerics. The final mean estimate.

References

Gerard, D., & Hoff, P. (2015). Adaptive Higher-order Spectral Estimators. arXiv preprint arXiv:1505.02114.

See also

update_c, update_lambda, get_c.